requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pgvector==0.2.4

# Vector Database
pinecone-client==3.0.2

# Cache & Queue
redis==5.0.1
celery==5.3.6

# LLM & Embeddings
openai==1.10.0
anthropic==0.39.0
groq==0.11.0
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0
python-json-logger==2.0.7

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
ruff==0.1.14
mypy==1.8.0

# Typing Extensions
types-redis==4.6.0.20240106
